ingestion/requirements.txt
# API Clients
spotipy==2.23.0
lyricsgenius==3.0.1
pylast==5.2.0

# GCP
google-cloud-storage==2.14.0
google-cloud-bigquery==3.14.0
google-cloud-secret-manager==2.17.0

# Data processing
pandas==2.1.4
numpy==1.26.2

# Utilities
python-dotenv==1.0.0
requests==2.31.0
tenacity==8.2.3
pydantic==2.5.3
python-json-logger==2.0.7

# Development
pytest==7.4.3
pytest-cov==4.1.0
black==23.12.1
flake8==7.0.0